+++ include/flydsl/Dialect/FlyROCDL/IR/CopyAtom.td
@@ -48,26 +48,25 @@ def FlyROCDL_CopyOpLdsReadTranspose : FlyROCDL_CopyOp<"CopyOpCDNA4LdsReadTranspo
 //===----------------------------------------------------------------------===//
 // CopyOp GFX1250 — N-D TDM (Tensor Data Mover) async Global<->LDS copy (1-5D).
 //
-// Whole-tile DMA. The atom carries the tile descriptor as runtime state. The
-// global operand of the copy_atom_call is a *shape/direction token only*: only
-// its layout (compile-time tile shape) and address space (load vs store) are
-// read — its pointer is unused; the base address comes from the `base` state.
-// `rank` (1-5) is the tensor/tile rank. Stateful fields (via fly.atom.set_value):
+// Whole-tile DMA. The global operand of the copy_atom_call supplies the base
+// pointer (its pointer, NOT atom state) and the compile-time tile shape (its
+// layout); its address space picks the direction (load vs store). `rank` (1-5) is
+// the tensor/tile rank. Stateful fields (via fly.atom.set_value):
 //   `workgroup_mask` (default 0): MCAST mask, ORed into GROUP1 config [15:0].
-//   `base`: global tile base pointer (advance it to move to the next tile).
 //   `extent_i` (default INT32_MAX = no clamp): per-dim tensor extent (tensor dim
 //     order, 0=outermost) for OOB handling; tensor_dim = max(0, extent - warp
 //     start) while the tile dims stay static, so a ragged tile is HW-guarded.
 //   `stride_i` (default = unset sentinel): per-dim tensor stride in elements for
-//     dims 0..rank-2 (innermost stride is assumed 1). Unset falls back to the
-//     tile memref's static layout stride.
+//     dims 0..rank-2 (innermost stride is assumed 1). Unset falls back to the tile
+//     memref's static layout stride; set it for a dynamic/true global stride that
+//     differs from the packed tile-internal stride.
 //   `imm_offset` (default 0): i64 byte offset added to base (K-loop tile bump).
 //===----------------------------------------------------------------------===//
 
-// TDM moves the whole N-D tile in one DMA, so it overrides `getCopyRank` (== rank)
+// TDM moves the whole N-D tile in one DMA, so it carries the WholeTileCopy trait
 // to keep the expand-copy lowering from decomposing the tiled memref per element.
 def FlyROCDL_CopyOpGFX1250TDM : FlyROCDL_StatefulCopyOp<"CopyOpGFX1250TDM", "gfx1250.tdm",
-    /*traits=*/[], /*copyOverrides=*/["getCopyRank"]> {
+    /*traits=*/[Fly_WholeTileCopyTrait]> {
   let parameters = (ins
     // Tensor/tile rank (1-5).
     "int32_t":$rank,
