+++ lib/Dialect/FlyROCDL/GFX1250/CopyAtom.cpp
@@ -98,42 +98,40 @@ constexpr int32_t kOuterStrideUnset = static_cast<int32_t>(0x80000000);
 
 } // namespace
 
-// Stateful: the atom carries the whole TDM N-D descriptor — workgroup_mask
-// (MCAST) plus the global tile descriptor. Struct slots: {mask, base,
-// extent_0..4 (i32, per-dim tensor extent for OOB), stride_0..3 (i64, per-dim
-// tensor stride in elements; innermost stride is assumed 1), imm_offset (i64)}.
-// See CopyAtom.td for field semantics.
+// Stateful: the atom carries the TDM N-D descriptor — workgroup_mask (MCAST) plus
+// the tile geometry, EXCEPT the global base, which comes from the copy_atom_call
+// operand pointer. Struct slots: {mask, extent_0..4 (i32, per-dim tensor extent
+// for OOB), stride_0..3 (i64, per-dim tensor stride in elements; innermost stride
+// is assumed 1), imm_offset (i64)}. See CopyAtom.td for field semantics.
 static constexpr unsigned kMaxTdmRank = 5;
 
 std::optional<unsigned> CopyOpGFX1250TDMType::getFieldIndex(AtomStateField field) {
   switch (field) {
   case AtomStateField::WorkgroupMask:
     return 0;
-  case AtomStateField::Base:
-    return 1;
   case AtomStateField::Extent0:
-    return 2;
+    return 1;
   case AtomStateField::Extent1:
-    return 3;
+    return 2;
   case AtomStateField::Extent2:
-    return 4;
+    return 3;
   case AtomStateField::Extent3:
-    return 5;
+    return 4;
   case AtomStateField::Extent4:
-    return 6;
+    return 5;
   case AtomStateField::Stride0:
-    return 7;
+    return 6;
   case AtomStateField::Stride1:
-    return 8;
+    return 7;
   case AtomStateField::Stride2:
-    return 9;
+    return 8;
   case AtomStateField::Stride3:
-    return 10;
+    return 9;
   // Byte offset added to the global base at lowering (i64, carry-safe). Lets a
   // K-loop advance the tile by bumping one scalar (imm_offset) instead of
   // re-deriving the base pointer. Default 0 folds away.
   case AtomStateField::ImmOffset:
-    return 11;
+    return 10;
   default:
     return std::nullopt;
   }
@@ -150,8 +148,7 @@ static AtomStateField strideField(unsigned i) {
 Type CopyOpGFX1250TDMType::getConvertedType(MLIRContext *ctx) const {
   auto i32 = IntegerType::get(ctx, 32);
   auto i64 = IntegerType::get(ctx, 64);
-  auto glbPtr = LLVM::LLVMPointerType::get(ctx, /*addrSpace=*/1);
-  SmallVector<Type> fields = {i32, glbPtr};
+  SmallVector<Type> fields = {i32};    // workgroup_mask
   fields.append(kMaxTdmRank, i32);     // extent_0..4
   fields.append(kMaxTdmRank - 1, i64); // stride_0..3
   fields.push_back(i64);               // imm_offset
@@ -169,11 +166,8 @@ Value CopyOpGFX1250TDMType::getDefaultState(OpBuilder &builder, Location loc) co
   Value zero = arith::ConstantIntOp::create(builder, loc, 0, 32);
   Value noClamp = arith::ConstantIntOp::create(builder, loc, 0x7FFFFFFF, 32);
   Value strideUnset = arith::ConstantIntOp::create(builder, loc, kOuterStrideUnset, 64);
-  Value nullBase =
-      LLVM::ZeroOp::create(builder, loc, LLVM::LLVMPointerType::get(ctx, /*addrSpace=*/1));
   Value zero64 = arith::ConstantIntOp::create(builder, loc, 0, 64);
   insert(AtomStateField::WorkgroupMask, zero);
-  insert(AtomStateField::Base, nullBase);
   for (unsigned i = 0; i < kMaxTdmRank; ++i)
     insert(extentField(i), noClamp);
   for (unsigned i = 0; i < kMaxTdmRank - 1; ++i)
@@ -195,10 +189,6 @@ Value CopyOpGFX1250TDMType::setAtomState(OpBuilder &builder, Location loc, Value
     return nullptr;
   return LLVM::InsertValueOp::create(builder, loc, atomStruct, fieldValue, ArrayRef<int64_t>{*idx});
 }
-
-// TDM moves the whole N-D tile in one DMA; its geometry lives in the rank-N
-// memref layout, so the expand-copy lowering must emit a single rank-N call.
-unsigned CopyOpGFX1250TDMType::getCopyRank() const { return static_cast<unsigned>(getRank()); }
 
 Attribute CopyOpGFX1250TDMType::getThrLayout() const { return FxLayout(FxC(1), FxC(1)); }
 
@@ -274,10 +264,13 @@ LogicalResult CopyOpGFX1250TDMType::emitAtomCall(OpBuilder &builder, Location lo
   if (!isLoad && !isStore)
     return failure();
 
-  // The global operand only marks the direction; its base pointer, extents, and
-  // strides come from the atom descriptor state. The tile shape is compile-time on
-  // the operand layout (tensor dim order: index 0 = outermost, rank-1 = innermost).
+  // The global operand supplies the base pointer and the compile-time tile shape;
+  // its per-dim extents (OOB) and strides come from the atom descriptor state (the
+  // tile view's static stride is the packed tile-internal stride, not necessarily
+  // the true global stride, which may be dynamic). Tensor dim order on the layout:
+  // index 0 = outermost, rank-1 = innermost.
   fly::MemRefType glbMemTy = isLoad ? srcMemTy : dstMemTy;
+  Value glbPtr = isLoad ? src : dst;
   Value ldsPtr = isLoad ? dst : src;
   int32_t rank = getRank();
 
@@ -339,8 +332,8 @@ LogicalResult CopyOpGFX1250TDMType::emitAtomCall(OpBuilder &builder, Location lo
   }
   strideElems[rank - 1] = arith::ConstantIntOp::create(builder, loc, 1, 64); // innermost contiguous
 
-  Value glbBasePtr = stateField(AtomStateField::Base);
-  Value glbBase = LLVM::PtrToIntOp::create(builder, loc, i64Ty, glbBasePtr);
+  // Global base = the copy_atom_call operand pointer (not atom state).
+  Value glbBase = LLVM::PtrToIntOp::create(builder, loc, i64Ty, glbPtr);
   // i64 byte offset (default 0) added to the base so a K-loop can advance the tile
   // by bumping one scalar; the carry into glb_hi is handled by the i64 split below.
   glbBase = arith::AddIOp::create(builder, loc, glbBase, stateField(AtomStateField::ImmOffset));
